Always show content while extracting
Prevents empty screen from getting shown by waiting
for Mercury to finish and then replacing the content.
