+++ capy/src/main/java/com/jocmp/capy/articles/ExtractedContent.kt
@@ -1,0 +1,14 @@
+package com.jocmp.capy.articles
+
+import com.jocmp.capy.common.Async
+
+data class ExtractedContent(
+    val requestShow: Boolean = false,
+    val value: Async<String?> = Async.Uninitialized,
+) {
+    val isLoading: Boolean
+        get() = requestShow && value is Async.Loading
+
+    val isComplete: Boolean
+        get() = requestShow && value is Async.Success
+}
