+++ capy/src/main/res/raw/stylesheet.css
@@ -7,12 +7,12 @@ body {
   font-family: sans-serif;
   word-wrap: break-word;
   margin: 0;
-  padding: 2rem 4rem;
+  padding: 2rem 1rem;
   background-color: var(--color-surface);
   color: var(--color-on-surface);
 
-  @media only screen and (max-width: 768px) {
-    padding: 2rem 1rem;
+  @media only screen and (min-width: 769px) {
+    padding: 2rem 4rem;
   }
 }
 
@@ -78,5 +78,4 @@ th {
 .article__body {
   line-height: 1.6rem;
 }
-
 /* #endregion BEM */
