+++ capy/src/main/res/raw/template.html
@@ -24,7 +24,12 @@
         <div>{{byline}}</div>
         <div>{{feed_name}}</div>
       </a>
-      <div id="article-body" class="article__body article__body--text-size-{{text_size}}">{{body}}</div>
+      <div class="article__body article__body--text-size-{{text_size}}">
+        <div id="article-body-content">
+          {{body}}
+        </div>
+        {{script}}
+      </div>
     </article>
   </body>
 </html>
